refactor: StationResponse 생성 책임 리팩토링
