+++ src/main/java/nextstep/subway/applicaion/LineService.java
@@ -10,7 +10,6 @@ import nextstep.subway.domain.Station;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -72,7 +71,7 @@ public class LineService {
         Station downStation = stationService.findById(sectionRequest.getDownStationId());
         Line line = lineRepository.findById(lineId)
                 .orElseThrow(IllegalArgumentException::new);
-                
+
         line.addSection(line, upStation, downStation, sectionRequest.getDistance());
     }
 
@@ -82,20 +81,9 @@ public class LineService {
                 line.getName(),
                 line.getColor(),
                 line.getDistance(),
-                createStationResponses(line)
+                StationResponse.createStationResponses(line)
         );
         return lineResponse;
-    }
-
-    private List<StationResponse> createStationResponses(Line line) {
-        if (line.isEmptySections()) {
-            return Collections.emptyList();
-        }
-        List<Station> stations = line.getStations();
-        List<StationResponse> collect = stations.stream()
-                .map(it -> stationService.createStationResponse(it))
-                .collect(Collectors.toList());
-        return collect;
     }
 
     @Transactional
